"user-ca-bundle" should be consider external regardless of proxy
MCO will observe the `user-ca-bundle` configmap regardless of whether
the proxy resource points at it or not. So recert should consider all
certs inside this configmap to be external.
